checking if file already exists before creating them

+++ my_utils/kaldi_preprocessing.py
@@ -8,21 +8,29 @@ from scenarios.folds_creation import load_dictionaries
 
 def write_kaldi_fold(fold_filepath, kaldi_dir, fold_type):
     dst = kaldi_dir / f"ID_{fold_type}.lst"
-    shutil.copyfile(src=fold_filepath, dst=dst)
+    if not dst.exists():
+        # Avoid copying the same file for both OMR and AMT models as it is the same
+        shutil.copyfile(src=fold_filepath, dst=dst)
 
 
 def write_kaldi_vocabulary(w2i_filepath, kaldi_dir):
     w2i, _ = load_dictionaries(w2i_filepath)
-    with open(kaldi_dir / "chars.lst", "w") as f:
-        for w in w2i.keys():
-            f.write(f"{w}\n")
+    kaldi_w2i_path = kaldi_dir / "chars.lst"
+    if not kaldi_w2i_path.exists():
+        # Avoid copying the same file for both OMR and AMT models as it is the same
+        with open(kaldi_w2i_path, "w") as f:
+            for w in w2i.keys():
+                f.write(f"{w}\n")
 
 
 def write_kaldi_groundtruth(kaldi_dir, id_labels, labels):
-    with open(kaldi_dir / "grnTruth.dat", "w") as f:
-        for id, y_true in zip(id_labels, labels):
-            id = pathlib.Path(id)
-            f.write(" ".join([str(id.stem)] + y_true) + "\n")
+    gt_path = kaldi_dir / "grnTruth.dat"
+    if not gt_path.exists():
+        # Avoid copying the same file for both OMR and AMT models as it is the same
+        with open(gt_path, "w") as f:
+            for id, y_true in zip(id_labels, labels):
+                id = pathlib.Path(id)
+                f.write(" ".join([str(id.stem)] + y_true) + "\n")
 
 
 def write_kaldi_confmat(kaldi_dir, fold_type, id_preds, preds, preds_len):
